# legion/operator/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a EDI server.",
        "title": "EDI API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/api/v1/model/training": {
            "get": {
                "description": "Get list of Model Trainings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ModelTraining"
                ],
                "summary": "Get list of Model Trainings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model id label",
                        "name": "com.epam.legion.model.id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "model version label",
                        "name": "com.epam.legion.model.version",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v1.MTResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a Model Training. Result is updated Model Training.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ModelTraining"
                ],
                "summary": "Update a Model Training",
                "parameters": [
                    {
                        "description": "Update a Model Training",
                        "name": "mt",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/v1.MTRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a Model Training. Result is created Model Training.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ModelTraining"
                ],
                "summary": "Create a Model Training",
                "parameters": [
                    {
                        "description": "Create a Model Training",
                        "name": "mt",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/v1.MTRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/v1.MTResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete list of  Model Training by labels",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ModelTraining"
                ],
                "summary": "Delete list of Model Trainings by labels",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model id label",
                        "name": "com.epam.legion.model.id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "model version label",
                        "name": "com.epam.legion.model.version",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    }
                }
            }
        },
        "/api/v1/model/training/{name}": {
            "get": {
                "description": "Get a Model Training by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ModelTraining"
                ],
                "summary": "Get a Model Training",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model Training name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/v1.MTResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    }
                }
            },
            "delete": {
                "description": "Get a Model Training by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ModelTraining"
                ],
                "summary": "Get a Model Training",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model Training name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "routes.HTTPResult": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Success of error message",
                    "type": "string"
                }
            }
        },
...
        "v1.MTRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "MT name",
                    "type": "string"
                },
                "spec": {
                    "description": "MT specification. It is the same as ModelTraining CRD specification",
                    "type": "object",
                    "$ref": "#/definitions/v1alpha1.ModelTrainingSpec"
                }
            }
        },
        "v1.MTResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "MT name",
                    "type": "string"
                },
                "spec": {
                    "description": "MT specification. It is the same as ModelTraining CRD specification",
                    "type": "object",
                    "$ref": "#/definitions/v1alpha1.ModelTrainingSpec"
                },
                "status": {
                    "description": "MT status. It is the same as ModelTraining CRD status",
                    "type": "object",
                    "$ref": "#/definitions/v1alpha1.ModelTrainingStatus"
                }
            }
        },
...
        "v1alpha1.ModelTrainingSpec": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "entrypoint": {
                    "description": "Model training file. It can be python\\bash script or jupiter notebook",
                    "type": "string"
                },
                "env": {
                    "description": "Custom environment variables that should be setted before entrypoint invocation.\nIn ENVname:value format",
                    "type": "object"
                },
                "hyperparameters": {
                    "description": "Model training hyperparameters in parameter:value format",
                    "type": "object"
                },
                "image": {
                    "description": "Train image",
                    "type": "string"
                },
                "modelFile": {
                    "description": "Model file",
                    "type": "string"
                },
                "reference": {
                    "description": "VCS Reference",
                    "type": "string"
                },
                "resources": {
                    "description": "Resources for model container\nThe same format like k8s uses for pod resources.",
                    "type": "string"
                },
                "toolchain": {
                    "description": "Type of toolchain. Currently supports only python.\n+kubebuilder:validation:Enum=python,jupyter",
                    "type": "string"
                },
                "vcsName": {
                    "description": "Name of VCSCredential resource. Must exists",
                    "type": "string"
                }
            }
        },
        "v1alpha1.ModelTrainingStatus": {
            "type": "object",
            "properties": {
                "exitCode": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "modelImage": {
                    "type": "string"
                },
                "podName": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "state": {
                    "description": "+kubebuilder:validation:Enum=scheduling,fetching-code,running,capturing,succeeded,failed,unknown",
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
    }
}